Enum solang_parser::pt::Loc

pub enum Loc {
    Builtin,
    CommandLine,
    Implicit,
    Codegen,
    File(usize, usize, usize),
}

A code location.

Variants

Builtin

Builtin

CommandLine

Command line

Implicit

Implicit

Codegen

Codegen

File(usize, usize, usize)

The file number, start offset and end offset in bytes of the source file.

Implementations

impl Loc

pub fn begin_range(&self) -> Self

Returns this location’s beginning range.

pub fn end_range(&self) -> Self

Returns this location’s end range.

pub fn file_no(&self) -> usize

Returns this location’s file number.

Panics

If this location is not a file.

pub fn try_file_no(&self) -> Option<usize>

Returns this location’s file number if it is a file, otherwise None.

pub fn start(&self) -> usize

Returns this location’s start.

Panics

If this location is not a file.

pub fn end(&self) -> usize

Returns this location’s end.

Panics

If this location is not a file.

pub fn exclusive_end(&self) -> usize

Returns this location’s end. The returned endpoint is not part of the range. This is used when a half-open range is needed.

Panics

If this location is not a file.

pub fn use_start_from(&mut self, other: &Loc)

Replaces this location’s start with other’s.

Panics

If this location is not a file.

pub fn use_end_from(&mut self, other: &Loc)

Replaces this location’s end with other’s.

Panics

If this location is not a file.

pub fn with_start_from(self, other: &Self) -> Self

See Loc::use_start_from.

Panics

If this location is not a file.

pub fn with_end_from(self, other: &Self) -> Self

See Loc::use_end_from.

Panics

If this location is not a file.

pub fn with_start(self, start: usize) -> Self

Creates a new Loc::File by replacing start.

Panics

If this location is not a file.

pub fn with_end(self, end: usize) -> Self

Creates a new Loc::File by replacing end.

Panics

If this location is not a file.

pub fn range(self) -> Range<usize>

Returns this location’s range.

Panics

If this location is not a file.

pub fn union(&mut self, other: &Self)

Performs the union of two locations